Assign seperate option for each test.
Previously, this PR added all test group values for all experiments to a
single array. This meant that multiple experiments could not be used on
a single page as this would result in a race.

This instead assigns a single option per experiment. This should be fine
initially as we wont be having overlapping experiments.